Override the primary GET http call, drop path property and drop support for azapi_resource since no use-case yet
